+++ obj_cyl_cone.c
@@ -1,26 +1,26 @@
 #include "rtv1.h"
 
-double	c2(double x)
+float	c2(float x)
 {
-	double	a;
+	float	a;
 
 	a = cos(x);
 	return (a * a);
 }
 
-double	s2(double x)
+float	s2(float x)
 {
-	double	a;
+	float	a;
 
 	a = sin(x);
 	return (a * a);
 }
 
-int		solv_equa(t_hit_equa *q, t_obj *s, double *t)
+int		solv_equa(t_hit_equa *q, t_obj *s, float *t)
 {
-	double	time;
-	double	d;
-	double	temp;
+	float	time;
+	float	d;
+	float	temp;
 
 	d = q->b * q->b - 4 * q->a * q->c;
 	if (d < 0)
@@ -44,11 +44,11 @@ int		solv_equa(t_hit_equa *q, t_obj *s, double *t)
 	return (1);
 }
 
-int		hit_cone(t_obj *s, t_vect *r_pos, t_vect *r_dir, double *t)
+int		hit_cone(t_obj *s, t_vect *r_pos, t_vect *r_dir, float *t)
 {
 	t_hit_equa	q;
-	double		x;
-	double		y;
+	float		x;
+	float		y;
 	t_vect		delta;
 
 	q.r_pos = r_pos;
@@ -73,7 +73,7 @@ int		hit_cone(t_obj *s, t_vect *r_pos, t_vect *r_dir, double *t)
 	return (1);
 }
 
-int		hit_cylinder(t_obj *s, t_vect *r_pos, t_vect *r_dir, double *t)
+int		hit_cylinder(t_obj *s, t_vect *r_pos, t_vect *r_dir, float *t)
 {
 	t_hit_equa	q;
